requirements.txt
# Core dependencies - NumPy 1.x compatible
numpy==1.26.4
scipy==1.11.4
numba==0.59.0

# Audio processing
openai-whisper==20231117
pyaudio==0.2.14

# PyTorch for M2/MPS
torch==2.1.0
torchaudio==2.1.0

# LLM & NLP
openai==1.6.1
langchain==0.1.0
langchain-community==0.0.10
transformers==4.36.0
sentence-transformers==2.2.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.0
colorama==0.4.6

# TTS
pyttsx3==2.90